+++ streamlit_antd_components/frontend/src/ts/Tabs.tsx
@@ -11,6 +11,7 @@ interface TabsProp {
     align: string;
     position: any;
     variant: any;
+    centered: boolean;
     height: number | null;
     use_container_width: boolean;
     size: any
@@ -22,11 +23,12 @@ interface TabsProp {
 
 const AntdTabs = (props: TabsProp) => {
     //get data
-    const items = strToNode(props['items'],props['size'])
+    const items = strToNode(props['items'])
     const index = props['index']
     const align = props['align']
     const position = props['position']
     const variant = props['variant']
+    const centered = props['centered']
     const height = props['height']
     const grow = props['use_container_width']
     const size = props['size']
@@ -79,7 +81,7 @@ const AntdTabs = (props: TabsProp) => {
             flex-grow: 1;
         }
     `
-    insertStyle('sac.tabs.style', grow ? style + growStyle : style)
+    insertStyle('sac.tabs-style', grow ? style + growStyle : style)
 
     //component height
     let tabsHeight = ['left', 'right'].includes(position) && height != null ? height : undefined
@@ -129,6 +131,8 @@ const AntdTabs = (props: TabsProp) => {
                         fontFamily: 'var(--font)',
                         cardBg: 'transparent',
                         cardGutter: variant === 'outline' ? 0 : 2,
+                        horizontalItemGutter: 15,
+                        horizontalMargin: '0',
                         cardHeight: getSize(size) + 25,
                         colorBorderSecondary: 'transparent',
                     },
@@ -142,6 +146,7 @@ const AntdTabs = (props: TabsProp) => {
                 onTabClick={onClick}
                 type={variant === 'default' ? 'line' : 'card'}
                 tabPosition={position}
+                centered={centered}
                 style={{height: tabsHeight}}
             />
         </ConfigProvider>
